Added a MarketActor...and realized that I know longer need a MarketLike trait.
