+++ src/main/scala-2.11/markets/MarketActor.scala
@@ -18,18 +18,36 @@ package markets
 import akka.actor.{ActorRef, Props, Actor}
 import markets.clearing.ClearingMechanismActor
 import markets.clearing.engines.MatchingEngineLike
+import markets.orders.OrderLike
 import markets.tradables.Tradable
 
 
+/** Actor for modeling markets.
+  *
+  * A `MarketActor` actor should directly receive `AskOrderLike` and `BidOrderLike` orders for a particular
+  * `Tradable` (filtering out any invalid orders) and then forward along all valid orders to a
+  * `ClearingMechanismActor` for further processing.
+  * @param matchingEngine The `MarketActor` uses the `matchingEngine` to construct a `ClearingMechanismActor`.
+  * @param settlementMechanism The `MarketActor` uses the `settlementMechanism` to construct a `ClearingMechanismActor`.
+  * @param tradable The object being traded on the market.
+  */
 class MarketActor(matchingEngine: MatchingEngineLike,
                   settlementMechanism: ActorRef,
-                  val tradable: Tradable) extends Actor
-  with MarketLike {
+                  val tradable: Tradable) extends Actor {
 
   /** Each `MarketActor` has a unique clearing mechanism. */
   val clearingMechanism: ActorRef = {
     context.actorOf(ClearingMechanismActor.props(matchingEngine, settlementMechanism),
       "clearing-mechanism")
+  }
+
+  def receive: Receive = {
+    case order: OrderLike if order.tradable == tradable =>
+      clearingMechanism forward order
+      sender() ! OrderAccepted
+    case order: OrderLike if !(order.tradable == tradable) =>
+      sender() ! OrderRejected
+    case _ => ???
   }
 
 }
